split cdf
